refactor(functions): consolidate the _axios shim into _shared
Thirteen byte-identical copies of _axios.ts become one at
plugins/functions/_shared/_axios.ts, with all 30 import sites rewritten to a
relative path into it. Every rewritten specifier was checked to resolve on disk.

This was the intended end state all along; the copies existed because the shim
replaced axios per function. Keeping them meant every future fix had 13x the
review surface and could silently be applied to one copy only -- which is how the
header and arraybuffer regressions reached review in the first place.

Also removes internal/scripts/check-axios-shim-sync.mjs, its npm script and its
workflow step. It existed solely to detect drift between the copies, and it pins
the old canonical path (alp-usermgmt/src/api/_axios.ts), so leaving it in place
would have failed the workflow setup job with exit 2.

Verified: single copy compiles and still passes the behaviour checks for header
bracket access, Buffer payloads and whole-response timeout; workflow YAML and
package.json still parse.
