+++ package.json
@@ -22,7 +22,6 @@
     "local": "npx cross-env ENV_TYPE=local CADDY__CONFIG=./deploy/caddy-config node ./scripts/dist/cli.js --version develop --demo -n .env.local --function-path ./plugins/functions -p 41100",
     "build:atlas": "bash scripts/build-atlas.sh",
     "build": "npm run build:atlas && npm run local -- build",
-    "check:axios-shim": "node internal/scripts/check-axios-shim-sync.mjs",
     "start": "npm run local -- start",
     "stop": "npm run local -- stop",
     "clean": "npm run local -- cleanci",
