+++ plugins/functions/alp-usermgmt/src/api/BaseIDPAPI.ts
@@ -1,4 +1,4 @@
-import { AxiosRequestConfig } from "./_axios.ts"
+import { AxiosRequestConfig } from "../../../_shared/_axios.ts"
 import { Service } from 'typedi'
 import https from 'https'
 import { createLogger } from '../Logger'
