+++ plugins/functions/alp-usermgmt/src/api/PortalAPI.ts
@@ -1,5 +1,5 @@
 import { Container, Service } from 'typedi'
-import { AxiosRequestConfig } from "./_axios.ts"
+import { AxiosRequestConfig } from "../../../_shared/_axios.ts"
 import { createLogger } from '../Logger'
 import { CONTAINER_KEY } from '../const'
 import https from 'https'
