+++ plugins/functions/concept-mapping/src/api/PortalAPI.ts
@@ -1,7 +1,7 @@
 import { env } from "../env";
 import { Dataset } from "../types";
 import https from "https";
-import { AxiosRequestConfig } from "./_axios.ts";
+import { AxiosRequestConfig } from "../../../_shared/_axios.ts";
 
 export class PortalAPI {
   private readonly baseURL: string;
