+++ plugins/functions/d2e-webapi/src/api/JobPluginsAPI.ts
@@ -1,4 +1,4 @@
-import { isAxiosError } from "./_axios.ts";
+import { isAxiosError } from "../../../_shared/_axios.ts";
 import { env } from "../env.ts";
 import { ICohortGeneratorFlowRun } from "./types.ts";
