+++ plugins/functions/dataset/api/DbCredentialsAPI.ts
@@ -1,5 +1,5 @@
 // import https from "node:https";
-import { AxiosRequestConfig } from "./_axios.ts";
+import { AxiosRequestConfig } from "../../_shared/_axios.ts";
 import { services } from "../env.ts";
 import { get, put } from "./request-util.ts";
