+++ plugins/functions/dataset/api/FhirGatewayAPI.ts
@@ -1,4 +1,4 @@
-import { AxiosRequestConfig } from "./_axios.ts";
+import { AxiosRequestConfig } from "../../_shared/_axios.ts";
 import { services } from "../env.ts";
 
 export class FhirGatewayAPI {
