+++ plugins/functions/dataset/api/JobpluginsAPI.ts
@@ -1,5 +1,5 @@
 // import https from "node:https";
-import { AxiosRequestConfig } from "./_axios.ts";
+import { AxiosRequestConfig } from "../../_shared/_axios.ts";
 import {
   ICreateDatamodelFlowRunDto
 } from "../../jobplugins/src/types.ts";
