+++ plugins/functions/jobplugins/src/api/AnalyticsAPI.ts
@@ -1,5 +1,5 @@
 import https from "node:https";
-import { AxiosRequestConfig } from "./_axios.ts";
+import { AxiosRequestConfig } from "../../../_shared/_axios.ts";
 import { env, services } from "../env.ts";
 
 export class AnalyticsSvcAPI {
