+++ plugins/functions/portal/src/analytics/analytics.api.ts
@@ -1,5 +1,5 @@
 import { Injectable, SCOPE } from "@danet/core";
-import axios, { AxiosRequestConfig } from "../_axios.ts";
+import axios, { AxiosRequestConfig } from "../../../_shared/_axios.ts";
 import { services } from "../env.ts";
 import {
   IDatabaseSchemaFilterResult,
