+++ plugins/functions/white-rabbit/src/api/filesManagerAPI.ts
@@ -1,4 +1,4 @@
-import { AxiosRequestConfig } from "./_axios.ts";
+import { AxiosRequestConfig } from "../../../_shared/_axios.ts";
 import { env } from "../env.ts";
 import { get } from "./request-util.ts";
